requirements.txt
Django==1.4.2
PIL==1.1.7
argparse==1.2.1
distribute==0.6.28
dj-database-url==0.2.1
django-class-based-auth-views==0.1a1
django-cuser==1.0
psycopg2==2.4.5
wsgiref==0.1.2
